# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ivmp-service API",
        "title": "ivmp-service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/swagger/ivmp-service",
    "paths": {
...
        "/camera": {
            "get": {
                "description": "query objects",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Camera"
                ],
                "summary": "query objects",
                "parameters": [
                    {
                        "type": "string",
                        "description": "_select",
                        "name": "_select",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "order",
                        "name": "_order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "identifier",
                        "name": "identifier",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "created_by",
                        "name": "created_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "created_time",
                        "name": "created_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "updated_by",
                        "name": "updated_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "updated_time",
                        "name": "updated_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ai_type",
                        "name": "ai_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "stream_type",
                        "name": "stream_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "stream_port",
                        "name": "stream_port",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "stream_path",
                        "name": "stream_path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "second_stream_path",
                        "name": "second_stream_path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ip",
                        "name": "ip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ai_model",
                        "name": "ai_model",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ai_status",
                        "name": "ai_status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ai_config",
                        "name": "ai_config",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "third_id",
                        "name": "third_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects array",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Camera"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "save",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Camera"
                ],
                "summary": "save",
                "parameters": [
                    {
                        "description": "object",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Camera"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Camera"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/camera/batch-delete": {
            "post": {
                "description": "batch delete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Camera"
                ],
                "summary": "batch delete",
                "parameters": [
                    {
                        "description": "id array",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/camera/batch-upsert": {
            "post": {
                "description": "batch update",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Camera"
                ],
                "summary": "batch update",
                "parameters": [
                    {
                        "description": "objects array",
                        "name": "entities",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/camera/groupby": {
            "get": {
                "description": "GroupBy, for example,  _select=level, then return  {level_val1:sum1,level_val2:sum2}, _where can input status=0",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Camera"
                ],
                "summary": "GroupBy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "_select",
                        "name": "_select",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "_where",
                        "name": "_where",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects array",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "additionalProperties": true
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
        "/camera/page": {
            "get": {
                "description": "page query, _page(from 1 begin), _page_size, _order, and others fields, status=1, name=$like.%CAM%",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Camera"
                ],
                "summary": "page query",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "current page",
                        "name": "_page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "_page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "order",
                        "name": "_order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "identifier",
                        "name": "identifier",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "created_by",
                        "name": "created_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "created_time",
                        "name": "created_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "updated_by",
                        "name": "updated_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "updated_time",
                        "name": "updated_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ai_type",
                        "name": "ai_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "stream_type",
                        "name": "stream_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "stream_port",
                        "name": "stream_port",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "stream_path",
                        "name": "stream_path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "second_stream_path",
                        "name": "second_stream_path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ip",
                        "name": "ip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ai_model",
                        "name": "ai_model",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ai_status",
                        "name": "ai_status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ai_config",
                        "name": "ai_config",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "third_id",
                        "name": "third_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects array",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/common.Page"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/model.Camera"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/camera/{id}": {
            "delete": {
                "description": "delete",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Camera"
                ],
                "summary": "delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "实例id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Camera"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.Page": {
            "type": "object",
            "properties": {
                "items": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "common.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "model.Camera": {
            "type": "object",
            "properties": {
                "ai_config": {
                    "description": "AI配置",
                    "type": "string"
                },
                "ai_model": {
                    "description": "AI模型",
                    "type": "string"
                },
                "ai_status": {
                    "description": "AI状态",
                    "type": "integer"
                },
                "ai_type": {
                    "description": "AI类型:(0:none,1:分类...)",
                    "type": "integer"
                },
                "created_by": {
                    "description": "创建者",
                    "type": "string"
                },
                "created_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "id",
                    "type": "string"
                },
                "identifier": {
                    "description": "标识",
                    "type": "string"
                },
                "ip": {
                    "description": "ip",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "second_stream_path": {
                    "description": "第二路流地址",
                    "type": "string"
                },
                "stream_path": {
                    "description": "流地址",
                    "type": "string"
                },
                "stream_port": {
                    "description": "流端口",
                    "type": "integer"
                },
                "stream_type": {
                    "description": "流类型(rtsp,rtmp,file)",
                    "type": "string"
                },
                "third_id": {
                    "description": "third_id",
                    "type": "string"
                },
                "type": {
                    "description": "类型(0:rtsp,1:virtual)",
                    "type": "integer"
                },
                "updated_by": {
                    "description": "更新者",
                    "type": "string"
                },
                "updated_time": {
                    "description": "更新时间",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
    }
}